pyproject.toml
[tool.poetry]
name = "cometml"
version = "0.1.0"
description = ""
authors = ["stefan"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
comet-ml = "^3.47.0"
langchain = "^0.3.3"
pandas = "^2.2.3"
langchain-community = "^0.3.2"
langchain-experimental = "^0.3.2"
langchain-openai = "^0.2.2"
graphdatascience = "^1.12"
tiktoken = "^0.8.0"
retry = "^0.9.2"
neo4j = "5.21.0"
sentence-transformers = "^3.2.0"
faiss-cpu = "^1.9.0"
python-dotenv = "^1.0.1"
langgraph = "^0.2.39"
langgraph-checkpoint-sqlite = "^2.0.0"
postgres = "^4.0"
streamlit = "^1.39.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
